+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+### Fixed
+* When meeting has ended normally, but not initiated by the client, send a meetingEnded event rather than a meetingFailed event.
+
 ## [0.26.2] - 2024-10-16
 
 ### Fixed